refactor, changed axios for fetch requests
